Major update: script add/update, cloud66 support
The laptop script is now safe to run at anytime, either when first
setting up or afterwards to update some tools. Not all tools are
updated on purpose. For example, major ones like any Homebrew installed
ones are needed to be updated manually as a precaution. Other tools are
safe to upload whenever. Some are just reinstalled every time the
script runs as its not critical and just easier to support/write (e.g.
cloud66, coda)