Use new link_dotfiles script instead
We now have dotfiles in this repo and is now symlinked and working fine
locally. Will be easier to manage these files together rather than
separately.

We had to source this file to get it to work. Need to learn more about
Bash as I thought I had to execute it using `./link_dotfiles` but that
causes an error as it seems BASE_PATH isn’t available. If I source the
file though it is available:

> source execute a shell script within the context of the current
shell. Since execution takes place within the context of the current
shell, any changes in the shell are retained following the completion
of the shell.
http://www.unix.com/2037-post2.html