Now using brew bundler for system deps and Mac apps
We run the default Brewfile with —no-upgrade flag since we want to be
explicit when we upgrade them since its uncommon. Even if we did they
are major and can cause other maintenance headaches (e.g. having to
recreate Postgres databases, etc.).

The custom Caskfile is for Mac apps but are safe to run even with
updates. There are just a few apps that we don’t want brew bundler to
update because they auto update.