Add 2 loggers to ActiveRecord to see failed queries
If a query fails with a StatementInvalid error, we'll see the error in the console and the message containing a portion of the query. If add a logger to a file, we'll see the full debug query with the actual parameter values. This helps to debug a query that fails in the console.