Add logging output helpers
Found this while reading through https://github.com/paveg/dotfiles and took the logging implementation but simplified to pass through shellcheck for /bin/sh interpreter.

This will help not just highlight with color but apply a timestamp which can be helpful if some of these scripts take a little while to run.