Install rbenv through Homebrew now
For a long time I've installed through git and its worked great but the preferred default installation is now through Homebrew. It all works really nicely and ruby-build is a dependency.

The ruby script now is much simpler as its just for installing a version of Ruby.

We'll make a similar change with Node to use nodenv instead which works and installs the same way.

Now uses the new logging helpers too.