Replace nvm with nodenv
This is much faster whereas nvm would take 1 second when opening new Terminal tabs since it seemed to be running node when doing so. Now that's gone with nodenv. The other benefit is that its essentially a copy of rbenv (and node-build for ruby-build) but for node so it has the same interface and works much the same way.

Now this node script is just for installing a node version with defaults.

The nodenv tool is installed through Homebrew, just like rbenv.