For ZSH, open Terminal tabs in the same directory
Apple has this implemented for Bash but since the default is changing in macOS Catalina to ZSH, we need to implement this ourselves to get that behavior. Once I upgrade to Catalina, perhaps Apple has implemented this already so it could be removed.

There are several different implementations of this but I went with what's here https://superuser.com/a/313172 as this also sets the title but updates for each command so its better at reflecting the same directory on a new tab.