Use .zshrc, not .zprofile
When I upgraded to Catalina I noticed my prompt wasn't working until I reloaded. I should have used an .zshrc file instead of a .zprofile file as it works fine now when opening a new terminal.

> There are many possible files where the zsh can load user configuration. You should use ~/.zshrc for your personal configurations.

https://scriptingosx.com/2019/06/moving-to-zsh-part-2-configuration-files/

This also removed code where we attempted to open terminal tabs in the same directory as its not needed in Catalina. This functionality already exists where as in Mojave it didn't for ZSH.