Add BUNDLER_EDITOR export to use VS Code
We need to set this so when running `bundle open gemName` it will use that editor to show the contents of the gem for editing.

In the development file, we add VS Code to our PATH so we can use its built in `code` command. We don't need to provide any arguments here as it'll know what to do.