Move any bash specific code to .profile
In the prompt, history and google files we were checking for bash and loading code specific to that but lets move it to .profile for now as its unlikely we'll switch back. If we need too, it'll be easier to update .profile to only source explicit files from the .profile.d instead.