On test failure, print all output from lines
If we display $output then everything is on one line but by using $lines instead we can display each line of output as it would have displayed outside the test. Tried this on the shellcheck.sh script since we will want to see the output if it does fail e.g. code warnings.