Run shellcheck as part of test suite
This now runs shellcheck against the code base as part of the test suite. We have to add an exit in case of failure so that the test suite will fail too and we can see the output to figure out what's wrong.

This exit code has no effect on running the script manually. It behaves the same way.