Install shellcheck so we can run it in CI
I thought the test suite in CI wasn't finding m shellcheck.sh file but that wasn't the problem. It wasn't finding shellcheck itself which makes sense as we have to install it. Anything until now we've been covered with Homebrew and somehow bats is already available to use on GitHub Actions.