Switch to zsh built in git completion
I'm going to give this a try this week as the former changes here would output a warning.

```
WARNING: this script is deprecated, please see git-completion.zsh
```

With this change, I had to address the prompt. Since the file to source remains the same and is specific to the prompt, its best to keep that code in the prompt file for now. https://git-scm.com/book/id/v2/Appendix-A%3A-Git-in-Other-Environments-Git-in-Zsh