Install yarn through npm globally
Noticed on the yarn docs that its not only suggested to install yarn through npm globally. This was not the case before as it was stated to avoid that approach. We now install globally through npm and can set the version per project. Overall prefer that instead of through brew.