Update installation instructions (settings sync)
The sync settings extension is deprecated because the feature is now built into Visual Studio Code. No need for a GitHub personal access token (PAT) for that which is deleted. We only need a PAT for Git CLI access.

Add helpful FAQ link for Xcode command line developer tools setup and managing it.

Add reminder to run `brew doctor` which is always helpful even when updating MacOS since it may show an issue with Xcode command line developer tools that requires reinstalling.