Update ruby/node setup scripts with version arg
These files were always meant to run standalone or with the laptop script. Since installing each has certain requirements (e.g. for Ruby we need to specify openssl path or default gems or actions to take) we want to run this for any version we install. Instead of having to manually edit this each time, the installation part is now in a function so the version is an argument. If this script is given an argument then it will install that version. The default is to install the latest version when running normally or as part of the laptop script.

With this change its important to set `set -e` so if for example an invalid or unrecognized version is provided, the rbenv install or nodenv install commands will exit and thus stop any further processing in this setup script. We don't want to run the next line(s) if the install command (or any other command for that matter) failed.