Update method for latest ruby/node version
For node script, simplify the version regex as the official Node releases are only listed in the "18.2.0" format while others have prefixes or suffixes.

For ruby script, use `--list` instead for stable only versions and a smaller list to parse. Further, where available use the full argument name `--invert-match` in place of the shorthand `-v` for clarity. Here we don't need to save characters since we aren't typing this out each time. It helps when looking at this to know what the argument is directly.