Fix issue with first time setup on new Mac
I noticed some errors or warnings that will require sourcing the zshrc profile. So for example, after installing Homebrew, the brew command wasn't available. When installing Ruby or Node for the first time, the command to set the global version didn't complete because of the same reason, zshrc needs to be reloaded.