Simplify ruby and node setup scripts
Update the scripts to require a single argument, the version. I used the no-argument approach as a default for installing the latest version. It would be easier to just set a specific default version for a global install and keep that up to date as needed. When running this script for the first time, it won't prompt since the version specified won't be installed. When running subsequent times, we'll get a nice prompted from rbenv and nodenv if we want to reinstall that version.

This will hopefully help with the last hurdle with running this on a new machine, where the rbenv global command wouldn't work despite reloading the profile.